pyproject.toml
[tool.poetry]
name = "Weather Aggregator"
version = "0.1.0"
description = ""
authors = ["SkanderBM <skander.bmahmoud@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2"
django-ninja = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
